apps/opsai-onboarding/agent-service/requirements.txt
# Core dependencies
crewai==0.1.3
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# LLM providers
openai==1.3.0
anthropic==0.7.0

# Vector stores
chromadb==0.4.18
pinecone-client==2.2.4

# Tools and utilities
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
celery==5.3.4
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Communication tools
twilio==8.10.0
sendgrid==6.10.0
stripe==7.8.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Development
pytest==7.4.3
black==23.11.0
mypy==1.7.1